+++ CHANGELOG.md
@@ -27,6 +27,7 @@ This build is not intended for submission, but rather for the developers to keep
   - Or they move to the wrong field.
 - On the player info, you can see how drunk the current player is.
 - A bug where the game crashed when you're drunk and the owner of the field is non-existant has been fixed.
+- Price and rent are now listed on the offers.
 - Income taxes got fixed.
 - Investments bought are assigned to the player
 - Game loop bug fixed when player lands on their own property